backend/requirements.txt
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
click==8.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
Django==3.2.9
djangorestframework==3.12.4
elasticsearch==7.15.2
entrypoints==0.3
fonttools==4.28.2
ipykernel==6.5.1
ipython==7.29.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
nltk==3.6.5
numpy==1.21.4
packaging==21.3
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
regex==2021.11.10
setuptools-scm==6.3.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
ElasticMock==1.8.1
django-cors-headers==3.10.1
